// src/context/Dashboard/types.ts
export enum ActionTypes {
    DASHBOARD_INIT_PENDING = 'DASHBOARD_INIT_PENDING',
    DASHBOARD_INIT_SUCCESS = 'DASHBOARD_INIT_SUCCESS',
    DASHBOARD_INIT_ERROR = 'DASHBOARD_INIT_ERROR',
    JWT_TOKEN_INIT_PENDING = 'JWT_TOKEN_INIT_PENDING',
    JWT_TOKEN_INIT_SUCCESS = 'JWT_TOKEN_INIT_SUCCESS',
    JWT_TOKEN_INIT_ERROR = 'JWT_TOKEN_INIT_ERROR',
    GET_FIELDS_INIT_PENDING = 'GET_FIELDS_INIT_PENDING',
    GET_FIELDS_INIT_SUCCESS = 'GET_FIELDS_INIT_SUCCESS',
    GET_FIELDS_INIT_ERROR = 'GET_FIELDS_INIT_ERROR',
    GET_ROWS_INIT_PENDING = 'GET_ROWS_INIT_PENDING',
    GET_ROWS_INIT_SUCCESS = 'GET_ROWS_INIT_SUCCESS',
    GET_ROWS_INIT_ERROR = 'GET_ROWS_INIT_ERROR',
    UPDATE_ROWS_PENDING = 'UPDATE_ROWS_PENDING',
    UPDATE_ROWS_SUCCESS = 'UPDATE_ROWS_SUCCESS',
    UPDATE_ROWS_ERROR = 'UPDATE_ROWS_ERROR',
    TABLE_UPDATE_SUCCESS = "TABLE_UPDATE_SUCCESS",
    TABLE_UPDATE_PENDING = 'TABLE_UPDATE_SUCCESS',
    TABLE_UPDATE_ERROR = 'TABLE_UPDATE_SUCCESS'
}
export type Action = {
    readonly type : ActionTypes;
    //eslint-disable-next-line @typescript-eslint/no-explicit-any
    readonly payload?:any;
}